requirements.txt
openai>=1.52
python-dotenv>=1.0
requests>=2.32
trafilatura>=1.9
beautifulsoup4>=4.12
playwright>=1.47
pydantic>=2.9
docxtpl>=0.16
simhash>=2.1.2
streamlit>=1.37
